Split timeline zoom/pan into secnd suite, to be extended
