+++ ui/e2e/timelines.spec.ts-snapshots/first-timeline-row-zoomed-chromium-darwin.png
